sourcemap-backend-main/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-cors-headers==4.9.0
langchain-google-genai==3.0.1
langchain-openai==1.0.2
langchain-community==0.4.1
pypdf==6.1.3
tenacity==9.1.2
sqlalchemy==2.0.44
pydantic==2.12.4
pydantic-settings==2.11.0
python-multipart==0.0.20
aiohttp==3.13.2
pytest==8.4.2
pytest-asyncio==1.2.0
drf-spectacular==0.29.0
loguru==0.7.2
langchain-pinecone==0.2.13
pandas==2.3.3
gunicorn==22.0.0
